// core/shared/service/enum/brand-settle-types.ts
/** 브랜드의 정산 정보 구분 */
export enum BrandSettleTypes {
  /** 국내 개인 */
  DOMESTIC_PERSONAL = 'DOMESTIC_PERSONAL',
  /** 국내 일반 과세자 */
  DOMESTIC_NORMAL_BUSINESS = 'DOMESTIC_NORMAL_BUSINESS',
  /** 국내 과세 특례 사업자 */
  DOMESTIC_SPECIAL_TAX_BUSINESS = 'DOMESTIC_SPECIAL_TAX_BUSINESS',
  /** 국내 간이 과세자(세금 계산서 발행 가능) */
  DOMESTIC_SIMPLE_WITH_INVOICE_BUSINESS = 'DOMESTIC_SIMPLE_WITH_INVOICE_BUSINESS',
  /** 국내 간이 과세자(세금 계산서 발행 불가능) */
  DOMESTIC_SIMPLE_NO_INVOICE_BUSINESS = 'DOMESTIC_SIMPLE_NO_INVOICE_BUSINESS',
  /** 국내 면세 사업자 */
  DOMESTIC_EXEMPT_BUSINESS = 'DOMESTIC_EXEMPT_BUSINESS',
  /** 국내 비영리 사업자 */
  DOMESTIC_NON_PROFIT_BUSINESS = 'DOMESTIC_NON_PROFIT_BUSINESS',
  /** 국내 고유 사업자 */
  DOMESTIC_UNIQUE_BUSINESS = 'DOMESTIC_UNIQUE_BUSINESS',
  /** 해외 */
  OVERSEA = 'OVERSEA',
}